fix: updating to mirage 131.0.0
BREAKING CHANGE: now requires mirage 131.0.0

+++ Assets/NetworkPositionSync/Runtime/SyncPositionBehaviour.cs
@@ -385,7 +385,7 @@ namespace JamesFrowen.PositionSync
                 {
                     payload = writer.ToArraySegment()
                 };
-                Client.Send(msg, (int)_system.MessageChannel);
+                Client.Send(msg, _system.MessageChannel);
             }
         }
